add NodeSubCommand + Backend
Add a separate subcommand list for for node operation related commands.
Right now they only include basic operations, but can be extended later
as per need.

Add a Backend struct that will hold the running bitcoind or electrsd
process in the background. The electrsd and bitcoind will be connected
together. And the wallet will connect to the backend electrsd via
electrum blockchain config. So in case of `regtest-electrum` too we will
operate the backend via rpc to the underlying core node.