feat(cbf): implement transaction broadcasting
The `allow(clippy...` is because the variable `txid` is immediately
returned, so binding can technically be omitted. I think removing the
variable assignment makes the code difficult to read, so I decided to
ignore the lint.

The actual implementation comes down to listening for an info message
that reports the transaction was sent to a peer. For simplicity I am
ignoring any wallet updates, but if the user calls the `Sync` command
they can catch them.