Update `bdk_kyoto` to `0.15.1`
Checkpoints were removed from Kyoto to minimize trust in the library and
reduce maintence burden. As such, there is no longer the ability to
"skip blocks", at least on test networks. Mainnet offers two
checkpoints, segwit and taproot activation, but these are included from
genesis on test networks.

Broadcasting a transaction now awaits for the transaction to be
gossiped, simplifying the logic here.