feat(silentpayments): add SilentPaymentCode command
The input for this command are two compressed public keys.

The network is obtained from the wallet current network.

The silent payment code generated is independent from any of the other
stateful features of bdk-cli.

This command is mainly intended for experimental use, do not lock any
funds to the generated code if you don't know what you are doing and
don't have the keys matchin the public keys used.